+++ plugins/filter_pages.ts
@@ -1,0 +1,30 @@
+import { merge } from "../core/utils.ts";
+
+import type { Extensions, Page, Site } from "../core.ts";
+
+export interface Options {
+  /** The list of extensions this plugin applies to */
+  extensions: Extensions;
+
+  /** The function to test the page */
+  fn: (page: Page) => boolean;
+}
+
+// Default options
+export const defaults: Options = {
+  extensions: "*",
+  fn: () => true,
+};
+
+/** A plugin to filter only some pages */
+export default function (userOptions?: Partial<Options>) {
+  const options = merge(defaults, userOptions);
+
+  return (site: Site) => {
+    site.process(options.extensions, (page: Page) => {
+      if (!options.fn(page)) {
+        return false;
+      }
+    });
+  };
+}